+++ src/main/java/baseball/domain/dto/BallNumbersDTO.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.dto;
 
 import java.util.ArrayList;
 import java.util.List;
